web/pyproject.toml
[tool.poetry]
name = "suitcase"
version = "0.1.0"
description = "Suitcase project"
authors = ["Yeham Kim <dev.yamkim@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
django = "^4.2.6"
djangorestframework = "^3.14.0"
requests = "^2.31.0"
celery = "^5.3.4"
django-filter = "^23.3"
django-compressor = "^4.4"
django-storages = "^1.14.2"
drf-spectacular = "^0.26.5"
django-celery-results = "^2.5.1"
channels = "^3.0.4"
django-eventstream = "^4.5.1"
psycopg2 = "^2.9.9"
pillow = "^10.1.0"
boto3 = "^1.28.78"
rembg = "^2.0.52"
django-redis = "^5.4.0"
django-stubs = "^4.2.6"
mypy = "^1.6.1"
djangorestframework-stubs = "^3.14.4"
celery-stubs = "^0.1.3"
celery-types = "^0.20.0"
types-pillow = "^10.1.0.0"

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[[tool.mypy.overrides]]
module = "rest_framework.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.development"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
